+++ utils/debug.py
@@ -1,5 +1,6 @@
 from . import addon
 
-def msg(*messages, space=0):
+
+def msg(*messages, spaces=0):
     if addon.debug():
-        print(*[str(element).ljust(space, ' ') for element in messages])
+        print(*[str(element).ljust(spaces, ' ') for element in messages])